# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca",
        "version": "pre-Alpha"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "JWTAuthenticator"
                ],
                "summary": "Registrar un Usuario desde JWT",
                "operationId": "5b1b8396e0f0688bc98f8183fce8e920",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Confirmar contraseña del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Con esta ruta se crea un nuevo usuario"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "JWTAuthenticator"
                ],
                "summary": "Logear un Usuario desde JWT",
                "operationId": "0e84dcf529e9de27079b2baa9301088c",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruta para hacer Login y crear un nuevo token"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/libros": {
            "get": {
                "tags": [
                    "libros"
                ],
                "summary": "Mostrar un listado de libros",
                "operationId": "5de8af456fc24b6be3c98332057468d3",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los libros."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "libros"
                ],
                "summary": "Añadir un Libro",
                "operationId": "e2798d77c0ea0393430855825b008022",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Título del libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "description": "Descripción del libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Con esta ruta se crea un nuevo libro"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/libros/{libros}": {
            "get": {
                "tags": [
                    "libros"
                ],
                "summary": "Mostrar un Libro por su id",
                "operationId": "18223e4cac7c0b9d8bc8470ddb41483f",
                "parameters": [
                    {
                        "name": "libros",
                        "in": "path",
                        "description": "Id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Con esta ruta se muestra un solo libro"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "libros"
                ],
                "summary": "Editar un Libro dada la id",
                "parameters": [
                    {
                        "name": "libros",
                        "in": "path",
                        "description": "Id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Título del libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "description": "Descripción del libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Con esta ruta se edita un libro"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "libros"
                ],
                "summary": "Borrar libro por su id",
                "operationId": "51693a7e5564c52bf525ccfd54567b72",
                "parameters": [
                    {
                        "name": "libros",
                        "in": "path",
                        "description": "Id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borra un libro con la id por parámetro"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}